requirement.txt
certifi==2022.9.24
charset-normalizer==2.1.1
elastic-transport==8.4.0
elasticsearch==8.5.1
idna==3.4
mysql-connector==2.2.9
numpy==1.23.5
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.12
